refactors and reorganization
